test(cordyceps): split list tests into modules (#227)
